requirements.txt
fastapi
uvicorn
requests
pillow
pytesseract
python-multipart
